requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
drf-yasg
python-decouple
